requirements/dev_thomas.txt
Django==1.5.4
South==0.8.2
Unipath==1.0
argparse==1.2.1
django-crispy-forms==1.4.0
django-debug-toolbar==0.10.2
django-extensions==1.2.5
gunicorn==18.0
psycopg2==2.5.1
pytz==2013.7
six==1.4.1
sqlparse==0.1.9
wsgiref==0.1.2
